now load torchscript modules directly but note that also provide metadata path
