+++ nbox/__init__.py
@@ -1,10 +1,12 @@
 import nbox.utils
 import nbox.model
 import nbox.load
+import nbox.parsers
 
 from nbox.model import Model
 from nbox.load import load, PRETRAINED_MODELS, PT_SOURCES
 from nbox.api import NBXApi
+from nbox.parsers import ImageParser, TextParser
 
 # from nbox.user import secrets
