+++ nbox/load.py
@@ -1,6 +1,8 @@
 # singluar loader file for all models in nbox
 
+import os
 import re
+import json
 from typing import Dict
 
 import inspect
@@ -9,6 +11,8 @@ import warnings
 from nbox.model import Model
 from nbox.api import NBXApi
 from nbox.utils import is_available
+
+import torch
 
 model_key_regex = re.compile(r"^((\w+)\/([\w\/-]+)):*([\w+:]+)?$")
 
@@ -202,26 +206,46 @@ def load(model_key: str = None, nbx_api_key: str = None, cloud_infer: bool = Fal
         nbox.Model: when using local inference
         nbox.NBXApi: when using cloud inference
     """
-    # the input key can also contain instructions on how to run a particular models and so
-    model_key_parts = re.findall(model_key_regex, model_key)
-    if not model_key_parts:
-        raise ValueError(f"Key: {model_key} incorrect, please check once!")
+    # step 1: check the model key if it is a file path, declare such a variable
+    if os.path.exists(model_key):
+        model_path = os.path.abspath(model_key)
+        model_meta_path = ".".join(model_path.split(".")[:-1] + ["json"])
+        assert os.path.exists(model_meta_path), f"Model meta file not found: {model_meta_path}"
 
-    # this key is valid, now get it's components
-    model_key, src, src_key, model_instr = model_key_parts[0]
-    if src not in PT_SOURCES:
-        raise ValueError(f"Model source: {src} not found. Is this package installed!")
-    model_fn, model_meta = PRETRAINED_MODELS.get(model_key, (None, None))
-    if model_meta is None:
-        model_fn, model_meta = PRETRAINED_MODELS.get(src, (None, None))
-        if model_meta is None:
-            raise IndexError(f"Model: {model_key} not found")
+        with open(model_meta_path, "r") as f:
+            model_meta = json.load(f)
+            spec = model_meta["spec"]
+        
+        category = spec["category"]
+        tokenizer = None
+        if category == "text":
+            from transformers import AutoTokenizer
+            tokenizer = AutoTokenizer.from_pretrained(model_path)
 
-    # load the model based on local infer or cloud infer
-    if cloud_infer and nbx_api_key:
-        out = NBXApi(model_key_or_url=model_key, category=model_meta, nbx_api_key=nbx_api_key)
+        model = torch.jit.load(model_path, map_location = "cpu")
+        out = Model(model, category=category, tokenizer=tokenizer, model_key=spec["model_key"], model_meta = model_meta)
+
     else:
-        model, model_kwargs = model_fn(model=src_key, model_instr=model_instr, **loader_kwargs)
-        out = Model(model=model, category=model_meta, model_key=model_key, **model_kwargs)
+        # the input key can also contain instructions on how to run a particular models and so
+        model_key_parts = re.findall(model_key_regex, model_key)
+        if not model_key_parts:
+            raise ValueError(f"Key: {model_key} incorrect, please check once!")
+
+        # this key is valid, now get it's components
+        model_key, src, src_key, model_instr = model_key_parts[0]
+        if src not in PT_SOURCES:
+            raise ValueError(f"Model source: {src} not found. Is this package installed!")
+        model_fn, model_meta = PRETRAINED_MODELS.get(model_key, (None, None))
+        if model_meta is None:
+            model_fn, model_meta = PRETRAINED_MODELS.get(src, (None, None))
+            if model_meta is None:
+                raise IndexError(f"Model: {model_key} not found")
+
+        # load the model based on local infer or cloud infer
+        if cloud_infer and nbx_api_key:
+            out = NBXApi(model_key_or_url=model_key, category=model_meta, nbx_api_key=nbx_api_key)
+        else:
+            model, model_kwargs = model_fn(model=src_key, model_instr=model_instr, **loader_kwargs)
+            out = Model(model=model, category=model_meta, model_key=model_key, model_meta = None, **model_kwargs)
 
     return out
