+++ nbox/utils.py
@@ -15,6 +15,9 @@ from rich.console import Console as richConsole
 
 import logging
 
+import numpy as np
+import torch
+
 # ----- functions
 
 logging.basicConfig(level="INFO")
@@ -73,6 +76,17 @@ def get_random_name():
 def hash_(item, fn="md5"):
     return getattr(hashlib, fn)(str(item).encode("utf-8")).hexdigest()
 
+
+def convert_to_list(x):
+  # recursively convert tensors -> list
+  if isinstance(x, list):
+    return x
+  if isinstance(x, dict):
+    return {k:convert_to_list(v) for k,v in x.items()}
+  elif isinstance(x, (torch.Tensor, np.ndarray)):
+    return x.tolist()
+  else:
+    raise Exception("Unknown type: {}".format(type(x)))
 
 # --- classes
