+++ pyproject.toml
@@ -20,12 +20,12 @@ documentation = "https://docs.nimblebox.ai/nbox/python-api"
 [tool.poetry.dependencies]
 python = "^3.6"
 rich = "10.7.0"
-randomname = "1.0.0"
+randomname = "0.1.3"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
 rich = "10.7.0"
-randomname = "1.0.0"
+randomname = "0.1.3"
 
 [build-system]
 requires = [
